refactor: replace Split component with SplitResizable for improved layout handling
