+++ src/components/split-resizable/index.tsx
@@ -1,0 +1,8 @@
+import '@/styles/split-resizable.css'
+import Split from 'react-split'
+import { SplitResizableProps, useSplitResizableService } from './use-service'
+
+export function SplitResizable(props: SplitResizableProps) {
+  const service = useSplitResizableService(props)
+  return <Split {...service.configs}>{props.children}</Split>
+}
