+++ src/styles/split-resizable.css
@@ -1,0 +1,3 @@
+.split-resizable .gutter {
+  cursor: ew-resize;
+}
